Added script for preprocessing BitPar grammar.
Includes two separate files for BitPar grammar rules and BitPar lexicon entries that are being read in.

NOTE: For some reason the NLTK grammar class does not allow grammar rules with additional/special characters which makes it not very useful for realistic applications. There's also no mechanism in NLTK to escape these additional characters. For this reason, it is probably necessary to include an own grammar implementation as well.