Added file for reading in the Wall Street Journal.
Since the BitPar grammar does not work with nltk.grammar because of occurring special characters in some of the rules, this script was added to demonstrate the CYK algorithm for data that can be read in via the nltk.grammar class. The script reads in the Wall Street Journal from nltk.corpus and creates a grammar for it that can be parsed by the CYK algorithm.